requirements.txt
streamlit
PyPDF2
python-dotenv
pydantic
vertexai
langgraph
typing-extensions
langchain
langchain-community
langchain-openai
beautifulsoup4
faiss-cpu
